adjustments for 1.7
